# source/repos/Restaurant-Reservation-Website-master/RestaurantReservation.Backend/RestaurantReservation.Api/Swaggerv1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Restaurant Reservation API v1",
    "description": "Restaurant Reservation API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Category/Count": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Retrive the total number of category in database.",
        "operationId": "GetCategoryCount",
        "responses": {
          "200": {
            "description": "Number of category record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving number of category record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/Category": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get all existing categories from the database.",
        "operationId": "GetAll",
        "responses": {
          "200": {
            "description": "A list of category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving categories info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Add a new category to the database. This checks for duplicate category name.",
        "operationId": "Add",
        "requestBody": {
          "description": "Category object",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A Boolean response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Error Adding a category object information into database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Retrieve a specific category information based on the Category ID given.",
        "operationId": "Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Category ID",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving a category info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "Update an existing category in the database. This checks for duplicate category name.",
        "operationId": "Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Category ID",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Category Object",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Category object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Category ID or Category name exists message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Category Version Conflict message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Updating a Category object information from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Category/{userId}/{id}": {
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Remove an existing category information from database.",
        "operationId": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Category ID",
              "format": "int64"
            }
          },
          {
            "name": "userID",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleting category successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Category record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Deleting a category object information from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/Schedule/Count/{restaurantID}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Retrive the total number of schedule of a restaurant in database.",
        "operationId": "GetCount",
        "parameters": [
          {
            "name": "restaurantID",
            "in": "path",
            "description": "Restaurant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Restaurant ID",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Number of schedule of a restaurant record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving number of schedule of a restaurant record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "categoryID": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 20,
            "type": "string"
          },
          "createdBy": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "editedBy": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "editedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "versionNo": {
            "type": "string",
            "nullable": true
          },
          "rowVersion": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: Bearer 12345abcdef",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}